# docs/swagger.yaml
openapi: 3.1.0
info:
  title: EMK Search Service
  version: 1.0.0
paths:
  /health:
    get:
      summary: Health
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /metrics:
    get:
      summary: Metrics
      operationId: metrics_metrics_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /build-index:
    post:
      summary: Build Index
      operationId: build_index_build_index_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/BuildIndexRequest'
              - type: 'null'
              title: Req
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /index-status:
    get:
      summary: Index Status
      operationId: index_status_index_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /suggest:
    post:
      summary: Suggest
      operationId: suggest_suggest_post
      requestBody:
        content:
          application/json:
            schema:
              type: string
              minLength: 1
              title: Query
              description: Tender item text to match against supplier catalogs
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Suggestion'
                type: array
                title: Response Suggest Suggest Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /catalogs:
    get:
      summary: Catalogs
      operationId: catalogs_catalogs_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /catalogs/upload:
    post:
      summary: Upload Catalog
      operationId: upload_catalog_catalogs_upload_post
      parameters:
      - name: rebuild_index
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Index
      - name: batch_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 256
          minimum: 1
          default: 16
          title: Batch Size
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_catalog_catalogs_upload_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /catalogs/upload-many:
    post:
      summary: Upload Many Catalogs
      operationId: upload_many_catalogs_catalogs_upload_many_post
      parameters:
      - name: rebuild_index
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Rebuild Index
      - name: batch_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 256
          minimum: 1
          default: 16
          title: Batch Size
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_many_catalogs_catalogs_upload_many_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /import-xlsx:
    post:
      summary: Import Xlsx
      operationId: import_xlsx_import_xlsx_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_xlsx_import_xlsx_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /export-xlsx:
    post:
      summary: Export Xlsx
      operationId: export_xlsx_export_xlsx_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Payload
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Body_import_xlsx_import_xlsx_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_import_xlsx_import_xlsx_post
    Body_upload_catalog_catalogs_upload_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
      type: object
      required:
      - file
      title: Body_upload_catalog_catalogs_upload_post
    Body_upload_many_catalogs_catalogs_upload_many_post:
      properties:
        files:
          items:
            type: string
            contentMediaType: application/octet-stream
          type: array
          title: Files
      type: object
      required:
      - files
      title: Body_upload_many_catalogs_catalogs_upload_many_post
    BuildIndexRequest:
      properties:
        batch_size:
          type: integer
          maximum: 256.0
          minimum: 1.0
          title: Batch Size
          default: 16
      type: object
      title: BuildIndexRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    Suggestion:
      properties:
        name:
          type: string
          title: Name
        price:
          anyOf:
          - type: number
          - type: 'null'
          title: Price
        similarity:
          type: number
          title: Similarity
        source_file:
          anyOf:
          - type: string
          - type: 'null'
          title: Source File
        sheet:
          anyOf:
          - type: string
          - type: 'null'
          title: Sheet
        row:
          anyOf:
          - type: integer
          - type: 'null'
          title: Row
      type: object
      required:
      - name
      - similarity
      title: Suggestion
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
servers:
- url: https://search.turbo-metallurg-montazh.ru
  description: Production search service
- url: http://localhost:8000
  description: Local development